+++ Classes/Domain/Repository/PaymentMethodRepository.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OliverKlee\Seminars\Domain\Repository;
+
+use OliverKlee\Seminars\Domain\Model\PaymentMethod;
+use TYPO3\CMS\Extbase\Persistence\Repository;
+
+/**
+ * @extends Repository<PaymentMethod>
+ */
+class PaymentMethodRepository extends Repository
+{
+}