+++ Tests/Functional/Domain/Repository/PaymentMethodRepositoryTest.php
@@ -1,0 +1,51 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OliverKlee\Seminars\Tests\Functional\Domain\Repository;
+
+use Nimut\TestingFramework\TestCase\FunctionalTestCase;
+use OliverKlee\Seminars\Domain\Model\PaymentMethod;
+use OliverKlee\Seminars\Domain\Repository\PaymentMethodRepository;
+use TYPO3\CMS\Core\Information\Typo3Version;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
+use TYPO3\CMS\Extbase\Object\ObjectManager;
+
+/**
+ * @covers \OliverKlee\Seminars\Domain\Model\PaymentMethod
+ * @covers \OliverKlee\Seminars\Domain\Repository\PaymentMethodRepository
+ */
+final class PaymentMethodRepositoryTest extends FunctionalTestCase
+{
+    protected $testExtensionsToLoad = ['typo3conf/ext/oelib', 'typo3conf/ext/seminars'];
+
+    /**
+     * @var PaymentMethodRepository
+     */
+    private $subject;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        if ((new Typo3Version())->getMajorVersion() >= 11) {
+            $this->subject = GeneralUtility::makeInstance(PaymentMethodRepository::class);
+        } else {
+            $objectManager = GeneralUtility::makeInstance(ObjectManager::class);
+            $this->subject = $objectManager->get(PaymentMethodRepository::class);
+        }
+    }
+
+    /**
+     * @test
+     */
+    public function mapsAllModelFields(): void
+    {
+        $this->importDataSet(__DIR__ . '/Fixtures/PaymentMethodRepository/PaymentMethodWithAllFields.xml');
+
+        $result = $this->subject->findByUid(1);
+
+        self::assertInstanceOf(PaymentMethod::class, $result);
+        self::assertSame('invoice', $result->getTitle());
+    }
+}